+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-* Fix OnLeave buttons handler.
+* Added option "Grab buttons without a name".
+* Minor fixes.